Cleanup / Bugfixes / Validation
