+++ gitopscli/abstract_git_util.py
@@ -1,30 +1,7 @@
 import os
-import sys
 from pathlib import Path
 from abc import ABC, abstractmethod
 from git import Repo
-from gitopscli.bitbucket_git_util import BitBucketGitUtil
-from gitopscli.github_git_util import GithubGitUtil
-
-
-def create_git(
-    username, password, git_user, git_email, organisation, repository_name, git_provider, git_provider_url, tmp_dir
-):
-    if git_provider == "bitbucket-server":
-        if not git_provider_url:
-            print(f"Please provide --git-provider-url for bitbucket-server", file=sys.stderr)
-            sys.exit(1)
-
-        git = BitBucketGitUtil(
-            tmp_dir, git_provider_url, organisation, repository_name, username, password, git_user, git_email
-        )
-    elif git_provider == "github":
-
-        git = GithubGitUtil(tmp_dir, organisation, repository_name, username, password, git_user, git_email)
-    else:
-        print(f"Git provider '{git_provider}' is not supported.", file=sys.stderr)
-        sys.exit(1)
-    return git
 
 
 class AbstractGitUtil(ABC):
