+++ gitopscli/apps_synchronizer.py
@@ -25,24 +25,31 @@ def sync_apps(apps_git, root_git):
     yaml = YAML()
     with open(bootstrap_values_file, "r") as stream:
         bootstrap = yaml.load(stream)
+    path_to_app_config = {}
+    all_other_apps = []
+    app_config_path = None
+    app_file_name = None
+    selected_app_config = None
     for boot_entry in bootstrap["bootstrap"]:
-        apps_config_file = root_git.get_full_file_path("apps/" + boot_entry["name"] + ".yaml")
-        path_to_app_config = {}
+        app_file_name = "apps/" + boot_entry["name"] + ".yaml"
+        apps_config_file = root_git.get_full_file_path(app_file_name)
         with open(apps_config_file, "r") as stream:
             app_config_content = yaml.load(stream)
-            path_to_app_config[str(apps_config_file)] = app_config_content
-    app_config_path = None
-    selected_app_config = None
-    for path, app_config in path_to_app_config.items():
-        print(app_config["repository"])
-        if app_config["repository"] == apps_git.get_clone_url():
-            selected_app_config = app_config
-            app_config_path = path
-            break
+            print(app_config_content["repository"])
+            if app_config_content["repository"] == apps_git.get_clone_url():
+                selected_app_config = app_config_content
+                app_config_path = str(apps_config_file)
+            else:
+                path_to_app_config[str(apps_config_file)] = app_config_content
+                all_other_apps += app_config_content["applications"].keys()
+    pprint.pprint(all_other_apps)
     if selected_app_config is None:
         raise Exception("Could't find config file with repository " + apps_git.get_clone_url() + " in apps/ directory")
+    for app_key in apps.keys():
+        if app_key in all_other_apps:
+            raise Exception("application: " + app_key + " already exists in a different repository")
 
     merge_yaml_element(app_config_path, "applications", apps, True)
 
-    root_git.commit("Updated applications in " + app_config_path)
+    root_git.commit("Updated applications in " + app_file_name)
     root_git.push("master")
